requirements.txt
streamlit
tensorflow
ultralytics
opencv-python
numpy
pillow
h5py
matplotlib
pandas
scikit-learn